MAY mark capture dates as NA
changed format_MAY based on information from data custodian.
Missing ring series: not available for now. Not sure how to deal with this.
Capture dates: not available either and perhaps better not to guess. It is unclear when these were done (could be during incubation but in publications it is written between d7-14 so may be a mix of both). This however makes the calc_age function not work.